charts/airbyte-chart/files/python/requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
pytest-mock==3.10.0
PyYAML==6.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
